ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
python-dotenv==1.0.0
httpx==0.25.2
openai==1.97.1
boto3==1.34.0
PyPDF2==3.0.1
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
asyncpg==0.29.0
redis==5.0.1